Use PncBuild to get proper version of bomArtifact. Refractor getConfiguredIndySettings file
