+++ pig/src/main/java/org/jboss/pnc/bacon/pig/impl/javadoc/JavadocManager.java
@@ -33,6 +33,7 @@ import org.jboss.pnc.bacon.pig.impl.pnc.ArtifactWrapper;
 import org.jboss.pnc.bacon.pig.impl.pnc.PncBuild;
 import org.jboss.pnc.bacon.pig.impl.utils.FileUtils;
 import org.jboss.pnc.bacon.pig.impl.utils.GAV;
+import org.jboss.pnc.bacon.pig.impl.utils.indy.Indy;
 import org.jboss.pnc.bacon.pig.impl.utils.pom.Dependency;
 import org.jboss.pnc.bacon.pig.impl.utils.pom.Profile;
 import org.jboss.pnc.bacon.pig.impl.utils.pom.Project;
@@ -143,7 +144,7 @@ public class JavadocManager extends DeliverableManager<GenerationData<?>, Void> 
 
     private void init() {
         this.temporaryDestination = FileUtils.mkTempDir("javadoc");
-        this.settingsXml = FileUtils.getConfiguredIndySettingsXmlPath(this.tempBuild);
+        this.settingsXml = Indy.getConfiguredIndySettingsXmlPath(this.tempBuild);
         this.localRepo = new File(temporaryDestination + File.separator + "localRepo");
         this.localRepo.mkdir();
         this.topLevelDirectory = new File(temporaryDestination, getTargetTopLevelDirectoryName());
