+++ pig/src/main/java/org/jboss/pnc/bacon/pig/impl/utils/FileUtils.java
@@ -37,7 +37,6 @@ import org.apache.commons.compress.compressors.xz.XZUtils;
 import org.apache.commons.compress.utils.IOUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.tools.ant.util.PermissionUtils;
-import org.jboss.pnc.bacon.pig.impl.utils.indy.Indy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -578,25 +577,14 @@ public final class FileUtils {
             String content = org.apache.commons.io.IOUtils.toString(fis, Charset.defaultCharset());
             content = content.replaceAll(orig, replace);
             FileOutputStream fos = new FileOutputStream(fileName);
-            org.apache.commons.io.IOUtils.write(content, new FileOutputStream(fileName), Charset.defaultCharset());
+            log.info("Replaced content");
+            org.apache.commons.io.IOUtils.write(content, fos, Charset.defaultCharset());
             fis.close();
             fos.close();
         } catch (IOException e) {
             log.error("Unable to change settings file", e);
+            throw new RuntimeException("Unable to change settings file " + e.getLocalizedMessage());
         }
     }
 
-    public static String getConfiguredIndySettingsXmlPath(boolean tempBuild) {
-        String settingsXml;
-        if (tempBuild) {
-            settingsXml = ResourceUtils.extractToTmpFile("/indy-temp-settings.xml", "settings", ".xml")
-                    .getAbsolutePath();
-
-        } else {
-            settingsXml = ResourceUtils.extractToTmpFile("/indy-settings.xml", "settings", ".xml").getAbsolutePath();
-        }
-        replaceFileString("\\$\\{INDY_URL}", Indy.getIndyUrl(), settingsXml);
-        replaceFileString("\\$\\{INDY_TMP_URL}", Indy.getIndyTempUrl(), settingsXml);
-        return settingsXml;
-    }
 }
